third_party/blink/web_tests/external/wpt/css/cssom/shorthand-values-expected.txt

This is a testharness.js-based test.
[FAIL] The serialization of border: 1px; border-top: 2px; should be canonical.
  assert_equals: expected "border-width: 2px 1px 1px; border-style: none; border-color: currentcolor; border-image: none;" but got "border-width: 2px 1px 1px; border-style: none; border-color: currentcolor; border-image: initial;"
[FAIL] The serialization of border: 1px; border-top: 1px !important; should be canonical.
  assert_equals: expected "border-right: 1px; border-bottom: 1px; border-left: 1px; border-image: none; border-top: 1px !important;" but got "border-right: 1px; border-bottom: 1px; border-left: 1px; border-image: initial; border-top: 1px !important;"
[FAIL] The serialization of border: 1px; border-top-color: red; should be canonical.
  assert_equals: expected "border-width: 1px; border-style: none; border-color: red currentcolor currentcolor; border-image: none;" but got "border-width: 1px; border-style: none; border-color: red currentcolor currentcolor; border-image: initial;"
[FAIL] The serialization of list-style-type: circle; list-style-position: inside; list-style-image: none; should be canonical.
  assert_equals: expected "list-style: inside circle;" but got "list-style: inside none circle;"
Harness: the test ran to completion.
